requirements.txt
Flask==2.2.5
PyYAML==6.0.1